Predict mimepowers (#38859)
* predict

* fix

* fix

* aa

* oops

* TOTALFIX

* some more cleanup

---------

(cherry picked from commit 4c24db9d9c137471cb5be9453c18e89d14fa286c)
