+++ Content.Server/Speech/Muting/MutingSystem.cs
@@ -1,4 +1,4 @@
-using Content.Server.Abilities.Mime;
+using Content.Shared.Abilities.Mime;
 using Content.Server.Chat.Systems;
 using Content.Server.Popups;
 using Content.Server.Speech.Components;
