+++ Resources/Locale/en-US/abilities/mime.ftl
@@ -1,5 +1,6 @@
 mime-cant-speak = Your vow of silence prevents you from speaking.
-mime-invisible-wall-popup = {CAPITALIZE(THE($mime))} brushes up against an invisible wall!
+mime-invisible-wall-popup-self = You brush up against an invisible wall!
+mime-invisible-wall-popup-others = {CAPITALIZE(THE($mime))} brushes up against an invisible wall!
 mime-invisible-wall-failed = You can't create an invisible wall there.
 mime-not-ready-repent = You aren't ready to repent for your broken vow yet.
 mime-ready-to-repent = You feel ready to take your vows again.
